spm: reduce spm_matcher size to uint8_t
No more space is needed.

+++ src/detect-content.c
@@ -913,7 +913,7 @@ static int DetectContentParseTest01 (void)
     const char *teststring = "abc\\:def";
     const char *teststringparsed = "abc:def";
 
-    uint16_t spm_matcher = SinglePatternMatchDefaultMatcher();
+    uint8_t spm_matcher = SinglePatternMatchDefaultMatcher();
     SpmGlobalThreadCtx *spm_global_thread_ctx = SpmInitGlobalThreadCtx(spm_matcher);
     FAIL_IF(spm_global_thread_ctx == NULL);
 
@@ -944,7 +944,7 @@ static int DetectContentParseTest02 (void)
     const char *teststring = "abc\\;def";
     const char *teststringparsed = "abc;def";
 
-    uint16_t spm_matcher = SinglePatternMatchDefaultMatcher();
+    uint8_t spm_matcher = SinglePatternMatchDefaultMatcher();
     SpmGlobalThreadCtx *spm_global_thread_ctx = SpmInitGlobalThreadCtx(spm_matcher);
     FAIL_IF(spm_global_thread_ctx == NULL);
 
@@ -975,7 +975,7 @@ static int DetectContentParseTest03 (void)
     const char *teststring = "abc\\\"def";
     const char *teststringparsed = "abc\"def";
 
-    uint16_t spm_matcher = SinglePatternMatchDefaultMatcher();
+    uint8_t spm_matcher = SinglePatternMatchDefaultMatcher();
     SpmGlobalThreadCtx *spm_global_thread_ctx = SpmInitGlobalThreadCtx(spm_matcher);
     FAIL_IF(spm_global_thread_ctx == NULL);
 
@@ -1006,7 +1006,7 @@ static int DetectContentParseTest04 (void)
     const char *teststring = "abc\\\\def";
     const char *teststringparsed = "abc\\def";
 
-    uint16_t spm_matcher = SinglePatternMatchDefaultMatcher();
+    uint8_t spm_matcher = SinglePatternMatchDefaultMatcher();
     SpmGlobalThreadCtx *spm_global_thread_ctx = SpmInitGlobalThreadCtx(spm_matcher);
     FAIL_IF(spm_global_thread_ctx == NULL);
 
@@ -1037,7 +1037,7 @@ static int DetectContentParseTest05 (void)
     DetectContentData *cd = NULL;
     const char *teststring = "abc\\def";
 
-    uint16_t spm_matcher = SinglePatternMatchDefaultMatcher();
+    uint8_t spm_matcher = SinglePatternMatchDefaultMatcher();
     SpmGlobalThreadCtx *spm_global_thread_ctx = SpmInitGlobalThreadCtx(spm_matcher);
     FAIL_IF(spm_global_thread_ctx == NULL);
 
@@ -1063,7 +1063,7 @@ static int DetectContentParseTest06 (void)
     const char *teststring = "a|42|c|44|e|46|";
     const char *teststringparsed = "abcdef";
 
-    uint16_t spm_matcher = SinglePatternMatchDefaultMatcher();
+    uint8_t spm_matcher = SinglePatternMatchDefaultMatcher();
     SpmGlobalThreadCtx *spm_global_thread_ctx = SpmInitGlobalThreadCtx(spm_matcher);
     FAIL_IF(spm_global_thread_ctx == NULL);
 
@@ -1094,7 +1094,7 @@ static int DetectContentParseTest07 (void)
     DetectContentData *cd = NULL;
     const char *teststring = "";
 
-    uint16_t spm_matcher = SinglePatternMatchDefaultMatcher();
+    uint8_t spm_matcher = SinglePatternMatchDefaultMatcher();
     SpmGlobalThreadCtx *spm_global_thread_ctx = SpmInitGlobalThreadCtx(spm_matcher);
     FAIL_IF(spm_global_thread_ctx == NULL);
 
@@ -1117,7 +1117,7 @@ static int DetectContentParseTest08 (void)
     DetectContentData *cd = NULL;
     const char *teststring = "";
 
-    uint16_t spm_matcher = SinglePatternMatchDefaultMatcher();
+    uint8_t spm_matcher = SinglePatternMatchDefaultMatcher();
     SpmGlobalThreadCtx *spm_global_thread_ctx = SpmInitGlobalThreadCtx(spm_matcher);
     FAIL_IF(spm_global_thread_ctx == NULL);
 
@@ -1407,7 +1407,7 @@ static int DetectContentParseTest09(void)
     DetectContentData *cd = NULL;
     const char *teststring = "boo";
 
-    uint16_t spm_matcher = SinglePatternMatchDefaultMatcher();
+    uint8_t spm_matcher = SinglePatternMatchDefaultMatcher();
     SpmGlobalThreadCtx *spm_global_thread_ctx = SpmInitGlobalThreadCtx(spm_matcher);
     FAIL_IF(spm_global_thread_ctx == NULL);
 
@@ -2072,7 +2072,7 @@ static int DetectContentParseTest41(void)
     }
     teststring[idx++] = '\0';
 
-    uint16_t spm_matcher = SinglePatternMatchDefaultMatcher();
+    uint8_t spm_matcher = SinglePatternMatchDefaultMatcher();
     SpmGlobalThreadCtx *spm_global_thread_ctx = SpmInitGlobalThreadCtx(spm_matcher);
     FAIL_IF(spm_global_thread_ctx == NULL);
 
@@ -2105,7 +2105,7 @@ static int DetectContentParseTest42(void)
     }
     teststring[idx++] = '\0';
 
-    uint16_t spm_matcher = SinglePatternMatchDefaultMatcher();
+    uint8_t spm_matcher = SinglePatternMatchDefaultMatcher();
     SpmGlobalThreadCtx *spm_global_thread_ctx = SpmInitGlobalThreadCtx(spm_matcher);
     FAIL_IF(spm_global_thread_ctx == NULL);
 
@@ -2139,7 +2139,7 @@ static int DetectContentParseTest43(void)
     }
     teststring[idx++] = '\0';
 
-    uint16_t spm_matcher = SinglePatternMatchDefaultMatcher();
+    uint8_t spm_matcher = SinglePatternMatchDefaultMatcher();
     SpmGlobalThreadCtx *spm_global_thread_ctx = SpmInitGlobalThreadCtx(spm_matcher);
     FAIL_IF(spm_global_thread_ctx == NULL);
 
@@ -2176,7 +2176,7 @@ static int DetectContentParseTest44(void)
     }
     teststring[idx++] = '\0';
 
-    uint16_t spm_matcher = SinglePatternMatchDefaultMatcher();
+    uint8_t spm_matcher = SinglePatternMatchDefaultMatcher();
     SpmGlobalThreadCtx *spm_global_thread_ctx = SpmInitGlobalThreadCtx(spm_matcher);
     FAIL_IF(spm_global_thread_ctx == NULL);
